refactor register text
